linux/node: reallocate nodeID upon conflict
NodeIDs and IPsec state suffer from a lack of reconciliation. If the
agent misses a node deletion event, stale state is never cleaned up.
This is somewhat known (#29822, #26298), but was generally considered
not of huge consequence. Stale XFRM states/policies can accumulate, but
will not match traffic - the effect is mostly slowing down processing
in agent and kernel. nodeIDs can eventually run out if too many node
deletions are missed, but the rate at which these are missed is expected
to be low.

Unfortunately, there are large clusters with high node churn in which
rare events become common, and hence the following sequence of events is
probable enough to actually observe:

1. a node is deleted while the agent is down (e.g. due to being
   upgraded)
2. a new node joins the cluster and is allocated IPs which overlap with
   previously used IPs.

If this occurs, the agent can have a partioned view of what nodeID this
node should have - in the BPF map, the k8s internal IP will map to a
different nodeID than the cilium internal ip. This breaks IPsec traffic
towards this node, as BPF applies a mark based on the BPF map nodeID of
the tunnnel endpoint, but the xfrm states expect to match the mark based
on the cilium internal IP. The result is traffic which doesn't match any
xfrm state/policy, falling back to the catch all block policy.

To work around this, we enforce that all IPs of a node get the same
nodeID - even if an IP was already pointing to an existing nodeID. Since
this node update is more current than whatever state we had held, it
seems more correct to ensure all IPs point to the same nodeID than
avoiding a BPF map write. We do so by forcing the allocation of a new
nodeID.

It is possible that unmapping the IPs but not deallocating the nodeID
can leak the nodeID. Deallocating unconditionally would be wrong, as the
stale mapping might point to a nodeID which represents a different,
alive node.

Signed-off-by: David Bimmler <david.bimmler@isovalent.com>