test: Apply nodeAffinity to all test components
Previously the node affinity was only applied to the Cilium Pods. As of
Cilium v1.16.0, other components like pre-flight would be scheduled onto
the node that is intended to _not_ run Cilium. This would then lead to
failures because there are Pods that never get scheduled / become ready.

Refactor the affinity configuration to apply it to each component used
in the testsuite.

Signed-off-by: Joe Stringer <joe@cilium.io>